backend/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-haystack==2.8.1
djangorestframework==3.9.0
elasticsearch==5.5.3
idna==2.8
pytz==2018.7
requests==2.21.0
six==1.12.0
urllib3==1.24.1
